Change the RegionInfo.RegionLocX/Y to meters. Also clean up the int vs uint by changing all instances to int instead of uint.

+++ OpenSim/Region/CoreModules/Framework/Monitoring/MonitorModule.cs
@@ -417,8 +417,8 @@ namespace OpenSim.Region.CoreModules.Framework.Monitoring
 
                     rb.ObjectCapacity = (uint)m_currentScene.RegionInfo.ObjectCapacity;
                     rb.RegionFlags = regionFlags;
-                    rb.RegionX = m_currentScene.RegionInfo.RegionLocX;
-                    rb.RegionY = m_currentScene.RegionInfo.RegionLocY;
+                    rb.RegionX = (uint)m_currentScene.RegionInfo.RegionLocX;
+                    rb.RegionY = (uint)m_currentScene.RegionInfo.RegionLocY;
 
                     ISimFrameMonitor simFrameMonitor = (ISimFrameMonitor)GetMonitor("SimFrameStats");
                     ITimeDilationMonitor timeDilationMonitor = (ITimeDilationMonitor)GetMonitor("Time Dilation");
